refactor: extract release assets to github/release_assets.py (#140)
